feat(conceal): `delete` command
