+++ src/bin/cnc/cli.rs
@@ -2,10 +2,10 @@ use std::path::PathBuf;
 
 use clap::Parser;
 
-/// Put entities into the recyle bin
+/// Put items into the recyle bin
 #[derive(Parser)]
 #[command(name = "cnc", version)]
 pub struct Cli {
-    /// The entities to be discarded
-    pub entities: Vec<PathBuf>,
+    /// The items to be discarded
+    pub items: Vec<PathBuf>,
 }
