--- src/handler/mod.rs
@@ -1,11 +1,0 @@
-#[cfg(any(freedesktop, target_os = "windows"))]
-mod clean;
-#[cfg(any(freedesktop, target_os = "windows"))]
-mod list;
-mod put;
-#[cfg(any(freedesktop, target_os = "windows"))]
-mod restore;
-
-pub use self::put::put;
-#[cfg(any(freedesktop, target_os = "windows"))]
-pub use self::{clean::clean, list::list, restore::restore};
