+++ src/util/mod.rs
@@ -1,0 +1,4 @@
+pub mod time;
+#[cfg(any(freedesktop, target_os = "windows"))]
+pub mod trash;
+pub mod tui;
